+++ imapinboxrules/exceptions/model/imap/ImapMailboxWithoutConnector.py
@@ -1,5 +1,0 @@
-
-class ImapMailboxWithoutConnector(Exception):
-
-  def __init__(self):
-    super().__init__("ImapMailbox was not initiated with connector.")