+++ CHANGELOG.md
@@ -6,9 +6,29 @@
 - [GitHub Action] Add badges in README.md (#5)
 - [GitHub Action] Disable commit count when merging on master (#6)
 - [GitHub Action] Block merge on master if there are "Merge commit" present (#9)
+- [CI] Add & run Lightweight IMAP server for tests
+- [CI] Add init script for setup lightweight IMAP server
+- [Exceptions] Rename `connector.ConnectorNotFoundException` into `connector.ConnectorNotFound`
+- [Exceptions] Create new exception `connector.ObjectWithoutConnector` when an object is initiated without connector and a method requiring it is called
+- [Exception] Create new exception `mailbox.MailboxNotFound` when selecting inexistant mailbox
+- [Exception] Create new exception `mailbox.MailboxCheckFailed` when mailbox check fail
+- [Exception] Remove exception `model.imap.ImapMailboxWithoutConnector` replaced by `connector.ObjectWithoutConnector`
+- [Exception] Create new exception `model.imap.ImapMailNotFound` when mail os not found
+- [Exception] Create new exception `model.imap.ImapMailNotLoaded` when mail is not loaded and a method requiring it is called
+- [ImapConnector] Allow to create connection with SSL-less security
+- [ImapConnector] Allow to select mailbox with method `ImapConnector.select_mailbox`
+- [ImapConnector] Allow to search for mail with method `ImapConnector.search_mail`
+- [ImapConnector] Allow to fetch mail with method `ImapConnector.fetch_mail`
+- [ImapMail] Create class to represent Mail object
+- [ImapMail] Allow to decode mail header and body
+- [ImapMailbox] Store mailbox full path in attribute `ImapMailbox.full_path`
+- [ImapMailbox] Use `ImapMailbox.full_path` attribute in `ImapMailbox.list_mailbox` method
+- [ImapMailbox] Create `ImapMailbox.select` method to select the current mailbox
+- [ImapMailbox] Add `ImapMailbox.count_mail` property to list email in mailbox
+- [ImapMailbox] Allow to search for mail with method `ImapConnector.search_mail`
 
 ### Bug fixes:
-- N/A
+- [ImapMailbox] Fix regex in `from_bytes` method when mailbox as no attributes
 
 ### Documentation:
 - Move documentation in [it's own directory](documentation/) (#7)